Added reporting feature to comments, added scrollview to login, updated provision profiles.

+++ AvanadeAIProject/Controllers/Detail/CommentDetailViewController.swift
@@ -21,9 +21,9 @@ class CommentDetailViewController: UIViewController {
         name.text = comment?.reviewer
         content.text = comment?.content
         if (comment?.editable)! {
-            deleteButton.isEnabled = true
+            deleteButton.title = "Delete"
         } else {
-            deleteButton.isEnabled = false
+            deleteButton.title = "Report"
         }
     }
     
@@ -33,22 +33,43 @@ class CommentDetailViewController: UIViewController {
         // Dispose of any resources that can be recreated.
     }
     @IBAction func onDeletePressed(_ sender: UIBarButtonItem) {
-        let alert = UIAlertController(title: nil, message: nil, preferredStyle: UIAlertControllerStyle.alert)
-        let attributedString = NSAttributedString(string: "Are you sure to delete this comment?", attributes: [
-            NSAttributedStringKey.font : UIFont(name: "Helvetica Neue", size: CGFloat(18))!
-            ])
+        if(comment?.editable)! {
+            let alert = UIAlertController(title: nil, message: nil, preferredStyle: UIAlertControllerStyle.alert)
+            let attributedString = NSAttributedString(string: "Are you sure to delete this comment?", attributes: [
+                NSAttributedStringKey.font : UIFont(name: "Helvetica Neue", size: CGFloat(18))!
+                ])
+            
+            alert.setValue(attributedString, forKey: "attributedTitle")
+            
+            alert.addAction(UIAlertAction(title: "Cancel", style: UIAlertActionStyle.cancel, handler: nil))
+            alert.addAction(UIAlertAction(title: "Confirm", style: UIAlertActionStyle.default, handler: { (action) in
+                DataManager.sharedInstance.deleteComment(comment: self.comment, completion: { (error) in
+                    if !self.handleError(error) {
+                        self.dismiss(animated: true, completion: nil)
+                    }
+                })
+            }))
+            present(alert, animated: true, completion: nil)
+        }
+        else {
+            let alert = UIAlertController(title: nil, message: nil, preferredStyle: UIAlertControllerStyle.alert)
+            let attributedString = NSAttributedString(string: "Are you sure to report this comment for inappropriate content?", attributes: [
+                NSAttributedStringKey.font : UIFont(name: "Helvetica Neue", size: CGFloat(18))!
+                ])
+            
+            alert.setValue(attributedString, forKey: "attributedTitle")
+            
+            alert.addAction(UIAlertAction(title: "Cancel", style: UIAlertActionStyle.cancel, handler: nil))
+            alert.addAction(UIAlertAction(title: "Confirm", style: UIAlertActionStyle.default, handler: { (action) in
+                DataManager.sharedInstance.reportComment(comment: self.comment, completion: { (error) in
+                    if !self.handleError(error) {
+                        self.dismiss(animated: true, completion: nil)
+                    }
+                })
+            }))
+            present(alert, animated: true, completion: nil)
+        }
         
-        alert.setValue(attributedString, forKey: "attributedTitle")
-
-        alert.addAction(UIAlertAction(title: "Cancel", style: UIAlertActionStyle.cancel, handler: nil))
-        alert.addAction(UIAlertAction(title: "Confirm", style: UIAlertActionStyle.default, handler: { (action) in
-            DataManager.sharedInstance.deleteComment(comment: self.comment, completion: { (error) in
-                if !self.handleError(error) {
-                    self.dismiss(animated: true, completion: nil)
-                }
-            })
-        }))
-        present(alert, animated: true, completion: nil)
     }
     
     @IBAction func onBackPressed(_ sender: UIBarButtonItem) {
